+++ src/models/index.ts
@@ -30,6 +30,7 @@ export * from './animal-onboarding';
 export * from './inter-tenant-access-contract';
 export * from './issue';
 export * from './jsonapi';
+export * from './linked-subject-identifier';
 export * from './jwe';
 export * from './jwk';
 export * from './jws';